# API/openapi-public.yaml
openapi: 3.0.3
info:
  title: Odyssey Finance API
  version: 1.1.0
  description: |
    오디세이 금융계획 웹 서비스의 외부 API (Spring). React 웹이 호출한다.

    ## 확정된 규약

    - **금액**: 원 단위 정수. 소수점 없음.
    - **비율**: 0~1 사이 소수. `0.15`가 15%다. 퍼센트 정수(`15`)를 보내면 안 된다.
    - **시각**: ISO 8601 + offset 필수. `2026-08-01T00:00:00+09:00` (O) / `2026-08-01` (X)
    - **날짜**: `date` 타입만 offset 없는 `YYYY-MM-DD` 사용 (생년월일, 목표일 등)
    - **월 표기**: `yearMonth`는 그 달 1일의 date. `2026-08-01` = 2026년 8월
    - **계획 생성**: 숫자·fan chart는 동기, LLM 설명만 Redis queue 후 별도 폴링

    ## 하이브리드 응답 방식

    `POST /goals/{goalId}/plan-versions`는 MC 시뮬레이션과 옵션 산출까지만 동기로
    처리하고 즉시 200을 반환한다. 정상 enqueue면 `explanation.status = PENDING`이고
    `explanation.text`는 null이다. Redis 발행이 실패하면 계획은 그대로 반환하고 설명만
    같은 응답에서 `FALLBACK`일 수 있다.

    LLM 설명은 백그라운드에서 생성되며, 클라이언트는
    `GET /plan-versions/{planVersionId}/explanation`을 2초 간격으로 폴링해
    `READY` 또는 `FALLBACK`이 될 때까지 기다린다. 설명이 실패해도 계획 자체는
    이미 사용 가능하다.

    ## 에러

    RFC 7807 `application/problem+json`. `code` 필드로 분기한다.
    409는 대부분 동시 요청 경합이므로, 클라이언트는 해당 리소스를 재조회하면 된다.
servers:
  - url: https://{host}/api/v1
    variables:
      host:
        default: localhost:8080
...
security:
  - bearerAuth: []
paths:
...
  /goals/{goalId}/plan-versions:
    parameters:
      - $ref: '#/components/parameters/GoalId'
    get:
      operationId: listPlanVersions
      tags: [plans]
      summary: 계획 버전 이력
      description: 재계획 이력 화면용. 최신 버전이 앞에 온다.
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [PROPOSED, ACTIVE, SUPERSEDED, REJECTED, INFEASIBLE, STALE]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PlanVersionSummary' }
    post:
      operationId: createPlanVersion
      tags: [plans]
      summary: 계획 생성 (하이브리드)
      description: |
        **동기 구간**: feasibility 사전검사 → MC 10,000 경로 → 70/80/90 옵션 산출
        → fan chart 분위수 저장. 수백 ms 안에 200을 반환한다.

        **비동기 구간**: LLM 설명 생성. 정상 enqueue면 응답의 `explanation.status`는
        `PENDING`이고 `explanation.text`는 null이다. Redis 발행 실패면 계산 결과를
        유지한 채 설명만 `FALLBACK`일 수 있다. `GET /plan-versions/{id}/explanation`을
        2초 간격으로 폴링할 것.

        기존에 미결정 `PROPOSED` 버전이 있으면 자동으로 `STALE` 처리되고
        새 버전이 유일한 `PROPOSED`가 된다.

        A(가용예산) < 0이면 MC를 돌리지 않고 `INFEASIBLE` 상태의 버전을 만든 뒤
        422를 반환한다. 목표금액 축소 · 기간 연장 · 소득 반영 중 하나가 필요하다.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                generationType:
                  type: string
                  enum: [INITIAL, USER_REQUESTED]
                  default: INITIAL
      responses:
        '200':
          description: 계획 생성됨 (설명은 PENDING 또는 Redis 장애 시 FALLBACK)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlanVersionDetail' }
        '409':
          description: |
            동시 요청 경합. 다른 요청이 이미 최신 제안을 만들었다.
            (code=PROPOSED_PLAN_EXISTS) 해당 goal의 계획을 재조회할 것.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '422':
          description: 수학적으로 목표 달성 불가 (code=PLAN_INFEASIBLE)
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - type: object
                    properties:
                      planVersionId: { type: integer }
                      infeasibleReason: { type: string }
                      shortfallAmount:
                        type: integer
                        format: int64
                        description: 유동지출을 0으로 해도 부족한 금액
  /plan-versions/{planVersionId}:
    parameters:
      - $ref: '#/components/parameters/PlanVersionId'
    get:
      operationId: getPlanVersion
      tags: [plans]
      summary: 계획 버전 상세 (옵션 + fan chart 포함)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlanVersionDetail' }
        '404':
          $ref: '#/components/responses/Problem404'
  /plan-versions/{planVersionId}/explanation:
    parameters:
      - $ref: '#/components/parameters/PlanVersionId'
    get:
      operationId: getPlanExplanation
      tags: [plans]
      summary: LLM 설명 폴링
      description: |
        하이브리드 응답의 두 번째 단계. 2초 간격 폴링을 권장하며,
        30초가 지나도 `PENDING`이면 폴링을 멈추고 설명 없이 계획만 보여줄 것.

        - `PENDING` — 생성 중
        - `PROCESSING` — worker가 작업을 가져가 생성 중
        - `READY` — LLM이 생성했고 숫자 검증(가드레일)을 통과
        - `FALLBACK` — timeout, Redis·Ollama 장애, 응답·숫자 검증 실패. 숫자 없는 템플릿으로 대체
        - `FAILED` — 생성 자체가 실패. 설명 영역을 숨길 것
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Explanation' }
        '404':
          $ref: '#/components/responses/Problem404'
  /plan-versions/{planVersionId}/custom-option:
    parameters:
      - $ref: '#/components/parameters/PlanVersionId'
    post:
      operationId: createCustomOption
      tags: [plans]
      summary: CUSTOM 옵션 생성 (사용자 baseline)
      description: |
        기획서 5-4. 제시된 70/80/90 계획이 과거 소비패턴 대비 과도하게 공격적일 때,
        사용자가 "현실적으로 유지 가능한 월 유동지출"을 직접 정하면 동일 시나리오
        집합에서 시뮬레이션 충족률을 재계산한다.

        **인과 순서 주의**: baseline을 먼저 정하고 그 결과로 충족률이 나온다.
        "86% 계획이므로 85만원" 같은 역순 표현을 UI에 쓰면 안 된다.

        한 계획 버전에 CUSTOM 옵션은 여러 개 만들 수 있으나, 실무상 마지막 것만
        보여주면 된다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [monthlySpending]
              properties:
                monthlySpending:
                  type: integer
                  format: int64
                  description: 사용자가 정한 월 유동지출 baseline (원)
      responses:
        '200':
          description: 생성됨
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlanOption' }
        '400':
          $ref: '#/components/responses/Problem400'
        '422':
          description: CUSTOM 금액이 확정 소비 하한보다 작음
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '409':
          description: 이 버전은 더 이상 선택 가능한 상태가 아님 (code=PLAN_NOT_SELECTABLE)
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /plan-versions/{planVersionId}/select-option:
    parameters:
      - $ref: '#/components/parameters/PlanVersionId'
    post:
      operationId: selectPlanOption
      tags: [plans]
      summary: 옵션 선택 및 계획 적용
      description: |
        선택한 옵션에 `selectedAt`을 기록하고 계획 버전을 `ACTIVE`로 전환한다.
        기존 ACTIVE 버전이 있으면 `SUPERSEDED`가 된다. 이 전부가 한 트랜잭션이다.

        이 요청 이후부터 shock · drift 트리거의 기준이 새 계획으로 재설정된다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [planOptionId]
              properties:
                planOptionId: { type: integer }
      responses:
        '200':
          description: 적용됨
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlanVersionDetail' }
        '404':
          $ref: '#/components/responses/Problem404'
        '409':
          description: |
            이미 다른 옵션이 선택됐거나(code=OPTION_ALREADY_SELECTED)
            버전 상태가 선택 불가(code=PLAN_NOT_SELECTABLE).
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }

  # ─────────────────────────────────────────── replan
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    GoalId:
      name: goalId
      in: path
      required: true
      schema: { type: integer }
    PlanVersionId:
      name: planVersionId
      in: path
      required: true
      schema: { type: integer }
  responses:
    Problem400:
      description: 요청 값 오류
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    Problem404:
      description: 없거나 접근 권한 없음
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
    Problem:
      type: object
      description: RFC 7807
      required: [type, title, status, detail, code, requestId]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        requestId: { type: string }
        code:
          type: string
          description: |
            클라이언트 분기용 코드.
            ACTIVE_GOAL_EXISTS · PROPOSED_PLAN_EXISTS · PLAN_INFEASIBLE ·
            PLAN_NOT_SELECTABLE · OPTION_ALREADY_SELECTED · DECISION_ALREADY_MADE ·
            ONBOARDING_INCOMPLETE · VALIDATION_FAILED
        errors:
          type: array
          description: 필드 단위 검증 오류
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
...
    PlanVersionSummary:
      type: object
      properties:
        id: { type: integer }
        versionNo: { type: integer }
        generationType:
          type: string
          enum: [INITIAL, MONTHLY_REGULAR, TRIGGERED_REPLAN, USER_REQUESTED]
        status:
          type: string
          enum: [PROPOSED, ACTIVE, SUPERSEDED, REJECTED, INFEASIBLE, STALE]
        asOfDate: { type: string, format: date }
        infeasibleReason: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        activatedAt: { type: string, format: date-time, nullable: true }
    PlanVersionDetail:
      allOf:
        - $ref: '#/components/schemas/PlanVersionSummary'
        - type: object
          properties:
            snapshot: { $ref: '#/components/schemas/PlanSnapshot' }
            options:
              type: array
              items: { $ref: '#/components/schemas/PlanOption' }
            explanation: { $ref: '#/components/schemas/Explanation' }
            simulation: { $ref: '#/components/schemas/SimulationInfo' }
    PlanSnapshot:
      type: object
      description: |
        계획 생성 당시의 값. 현재 값이 바뀌어도 과거 계획이 왜 그렇게
        계산됐는지 설명하기 위해 보존한다.
      properties:
        monthlyIncome: { type: integer, format: int64 }
        monthlyFixedCost: { type: integer, format: int64 }
        targetAmount: { type: integer, format: int64 }
        currentSaved: { type: integer, format: int64 }
        targetDate: { type: string, format: date }
        availableVariableBudget:
          type: integer
          format: int64
          description: A. 목표 기간 동안 유동지출에 쓸 수 있는 총액
        currentAvgVariableSpending: { type: integer, format: int64 }
        remainingMonths: { type: integer }
        resolvedSpendingFloor: { $ref: '#/components/schemas/ResolvedSpendingFloor' }
    PlanOption:
      type: object
      properties:
        id: { type: integer }
        optionType: { type: string, enum: [PRESET, CUSTOM] }
        nominalLevel:
          type: number
          nullable: true
          description: PRESET의 0.70 / 0.80 / 0.90. CUSTOM은 null
        recommendedMonthlySpending:
          type: integer
          format: int64
          description: 추천 월 유동지출 한도. 화면의 핵심 숫자
        requiredReductionRate:
          type: number
          description: 현재 평균 대비 필요 감축률. 소득 증가 시 음수 가능
        simulationCoverage:
          type: number
          description: |
            시뮬레이션 충족률 (0~1). **목표 달성 확률이 아니다.**
            UI에는 "시뮬레이션 충족률" 또는 "계획 안정성 수준"으로 표기할 것.
            "달성 확률 90%" 같은 확정적 표현은 사용 금지
        historicalFeasibilityRatio:
          type: number
          description: 최근 24개월 중 이 수준 이하로 소비한 달의 비율
        aggressiveWarning:
          type: boolean
          description: |
            true면 "이 계획은 최근 소비패턴보다 상당히 낮은 수준입니다" 경고와
            함께 소비내역 확인 흐름을 제안한다. 차단이 아니라 재확인 유도
        effectiveMaxReductionRate: { type: number, minimum: 0, maximum: 1 }
        floorApplied: { type: boolean }
        targetCoverageMet: { type: boolean }
        selectedAt: { type: string, format: date-time, nullable: true }
        percentileBands:
          type: array
          description: fan chart 재현용 월별 분위수
          items: { $ref: '#/components/schemas/PercentileBand' }
    PercentileBand:
      type: object
      properties:
        monthIndex: { type: integer, description: 계획 시작 후 N개월차 (1부터) }
        metricType: { type: string, example: CUMULATIVE_SAVINGS }
        p10: { type: integer, format: int64 }
        p25: { type: integer, format: int64 }
        p50: { type: integer, format: int64 }
        p75: { type: integer, format: int64 }
        p90: { type: integer, format: int64 }
    Explanation:
      type: object
      properties:
        status:
          type: string
          enum: [PENDING, PROCESSING, READY, FALLBACK, FAILED]
        text: { type: string, nullable: true }
        model: { type: string, nullable: true }
        generatedAt: { type: string, format: date-time, nullable: true }
        retryCount:
          type: integer
          nullable: true
          description: 가드레일 재시도 횟수 (운영 관찰용)
        promptVersion: { type: string, nullable: true }
        failedNumbers:
          type: array
          items: { type: string }
    ResolvedSpendingFloor:
      type: object
      required: [mode, requestedMonthlyAmount, effectiveMonthlyAmount]
      properties:
        mode: { type: string, enum: [OFF, AUTO, CUSTOM] }
        requestedMonthlyAmount: { type: integer, format: int64, minimum: 0 }
        effectiveMonthlyAmount: { type: integer, format: int64, minimum: 0 }
        autoHistoryMonths: { type: integer, nullable: true }
    SimulationInfo:
      type: object
      nullable: true
      description: INFEASIBLE 버전은 null
      properties:
        method: { type: string, example: IID_BOOTSTRAP }
        nPaths: { type: integer, example: 10000 }
        engineVersion: { type: string }
        createdAt: { type: string, format: date-time }
...